@@ -16,31 +16,42 @@ catppuccin-macchiato = "catppuccin.extras.pygments:MacchiatoStyle"
 catppuccin-mocha = "catppuccin.extras.pygments:MochaStyle"
 
 [tool.poetry.dependencies]
-python = "^3.7.2"
-pygments = { version = "^2.13.0", optional = true }
-rich = { version = "^13.3.5", optional = true }
+python = "^3.8.0"
+pygments = { version = "^2.17.2", optional = true }
+rich = { version = "^13.7.0", optional = true }
 
 [tool.poetry.extras]
 pygments = ["pygments"]
 rich = ["rich"]
 
 [tool.poetry.group.dev.dependencies]
-pylint = "^2.15.5"
-mypy = "^0.982"
-black = "^22.10.0"
-isort = "^5.10.1"
-pytest = "^7.2.0"
-pytest-cov = "^4.0.0"
-types-setuptools = "^65.5.0.3"
-types-pygments = "^2.13.1.1"
+mypy = "^1.8.0"
+pytest = "^8.0.0"
+pytest-cov = "^4.1.0"
+ruff = "^0.2.1"
+types-pygments = "^2.17.0.20240106"
+types-colorama = "^0.4.15.20240205"
 
 [tool.poetry-dynamic-versioning]
 enable = true
 
-[tool.pylint.messages_control]
-disable = [
-    "useless-import-alias", # pyright compatibility
+[tool.ruff.lint]
+select = ["ALL"]
+ignore = [
+  "D203",   # one-blank-line-before class, incompatible with D211
+  "D213",   # multi-line-summary-second-line, incompatible with D212
+  "ANN101", # missing-type-self, deprecated
+  "ANN102", # missing-type-cls, deprecated
+  "ISC001", # single-line-implicit-string-concatenation, incompatible with formatter
+  "COM812", # missing-trailing-comma, incompatible with formatter
 ]
 
-[tool.isort]
-profile = "black"
+[tool.ruff.lint.per-file-ignores]
+"tests/test_*.py" = [
+  "D100", # undocumented-public-module
+  "D103", # undocumented-public-function
+  "S101", # assert
+]
+
+[tool.mypy]
+strict = true
