@@ -1,13 +1,8 @@
-"""
-Utilities for working with HSL colours.
-"""
+"""Utilities for working with HSL colours."""
 from __future__ import annotations
 
-from typing import Tuple
 
-
-# pylint: disable=invalid-name
-def rgb_to_hsl(r: int, g: int, b: int) -> Tuple[int, float, float]:
+def rgb_to_hsl(r: int, g: int, b: int) -> tuple[int, float, float]:
     """Convert RGB to HSL."""
     r1 = r / 255
     g1 = g / 255
@@ -18,23 +13,19 @@ def rgb_to_hsl(r: int, g: int, b: int) -> Tuple[int, float, float]:
     delta = cmax - cmin
 
     if delta == 0:
-        h = 0.0
+        hue = 0.0
     elif cmax == r1:
-        h = ((g1 - b1) / delta) % 6
+        hue = ((g1 - b1) / delta) % 6
     elif cmax == g1:
-        h = (b1 - r1) / delta + 2
+        hue = (b1 - r1) / delta + 2
     else:
-        h = (r1 - g1) / delta + 4
+        hue = (r1 - g1) / delta + 4
 
-    h = round(h * 60)
-    if h < 0:
-        h += 360
+    hue = round(hue * 60)
+    if hue < 0:
+        hue += 360
 
-    l = (cmax + cmin) / 2
+    lightness = (cmax + cmin) / 2
+    saturation = 0 if delta == 0 else delta / (1 - abs(2 * lightness - 1))
 
-    if delta == 0:
-        s = 0.0
-    else:
-        s = delta / (1 - abs(2 * l - 1))
-
-    return h, round(s, 2), round(l, 2)
+    return hue, round(saturation, 2), round(lightness, 2)
