@@ -1,3 +1,5 @@
 """🐍 Soothing pastel theme for Python."""
-from catppuccin.colour import Colour as Colour
-from catppuccin.flavour import Flavour as Flavour
+from catppuccin.colour import Colour
+from catppuccin.flavour import Flavour
+
+__all__ = ["Colour", "Flavour"]
